+++ src/data/store/key/set/mod.rs
@@ -1,7 +1,14 @@
 // devela/src/data/store/key/set/mod.rs
 //
-//! Keyed set structures.
+#![doc = crate::_DOC_DATA_STORE_KEY_SET!()] // public
+#![doc = crate::_doc!(modules: crate::data::store::key; set)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: collections)]
 //
+
+#[cfg(feature = "alloc")]
+mod _reexport_alloc;
+mod _reexport_dep;
 
 mod sparse; // SparseSet[Array|Error]
 #[cfg(all(target_os = "linux", feature = "_linux_abi", feature = "unsafe_ffi"))]
@@ -12,5 +19,10 @@ crate::structural_mods! { // _mods
         pub use super::sparse::*;
         #[cfg(all(target_os = "linux", feature = "_linux_abi", feature = "unsafe_ffi"))]
         pub use super::linux_sparse::*;
+    }
+    _reexports {
+        #[cfg(feature = "alloc")]
+        pub use super::_reexport_alloc::*;
+        pub use super::_reexport_dep::*;
     }
 }